Scores are floating point. How did you not remember that.

+++ fed/src/event_schema.rs
@@ -313,9 +313,9 @@ pub enum FedEventData {
         game: GameEvent,
         winner_id: Uuid,
         winning_team_name: String,
-        winning_team_score: i32,
+        winning_team_score: f32,
         losing_team_name: String,
-        losing_team_score: i32,
+        losing_team_score: f32,
     },
 }
